警告全消し、警告をエラーとして扱うように

+++ engine/math/Matrix4.cpp
@@ -344,5 +344,5 @@ Vector3 operator*(const Vector3& v, const Matrix4& m)
 }
 
 float Matrix4::ConvertToRadian(float rad) {
-	return rad * (3.141592 / 180);
+	return rad * ((float)3.141592 / 180);
 }
